feat: provide service explanation component
