+++ src/components/registration/WelcomeScreen.tsx
@@ -1,13 +1,10 @@
 import * as React from 'react';
-import { ReactComponent as PenIcon } from '../../resources/img/icons/pen.svg';
-import { ReactComponent as EnvelopeIcon } from '../../resources/img/icons/envelope.svg';
-import { ReactComponent as SpeechBubbleIcon } from '../../resources/img/icons/speech-bubble.svg';
-import { ReactComponent as LockIcon } from '../../resources/img/icons/lock.svg';
 import { translate } from '../../utils/translate';
 import { Button, ButtonItem, BUTTON_TYPES } from '../button/Button';
 import { Text } from '../text/Text';
 import { config } from '../../resources/scripts/config';
 import { Headline } from '../headline/Headline';
+import { ServiceExplanation } from '../serviceExplanation/ServiceExplanation';
 import './welcomeScreen.styles';
 
 interface WelcomeScreenProps {
@@ -16,29 +13,6 @@ interface WelcomeScreenProps {
 }
 
 export const WelcomeScreen = (props: WelcomeScreenProps) => {
-	const welcomeScreenData = [
-		{
-			icon: <PenIcon />,
-			title: translate('registration.welcomeScreen.info1.title'),
-			text: translate('registration.welcomeScreen.info1.text')
-		},
-		{
-			icon: <EnvelopeIcon />,
-			title: translate('registration.welcomeScreen.info2.title'),
-			text: translate('registration.welcomeScreen.info2.text')
-		},
-		{
-			icon: <SpeechBubbleIcon />,
-			title: translate('registration.welcomeScreen.info3.title'),
-			text: translate('registration.welcomeScreen.info3.text')
-		},
-		{
-			icon: <LockIcon />,
-			title: translate('registration.welcomeScreen.info4.title'),
-			text: translate('registration.welcomeScreen.info4.text')
-		}
-	];
-
 	const registrationButton: ButtonItem = {
 		label: translate('registration.welcomeScreen.register.buttonLabel'),
 		type: BUTTON_TYPES.PRIMARY
@@ -53,21 +27,7 @@ export const WelcomeScreen = (props: WelcomeScreenProps) => {
 		<div className="registrationWelcome">
 			<Headline text={props.title} semanticLevel="2" />
 			<h4>{translate('registration.welcomeScreen.subline')}</h4>
-
-			<div className="registrationWelcome__infoWrapper">
-				{welcomeScreenData.map((infoItem, key) => (
-					<div className="registrationWelcome__infoItem" key={key}>
-						{infoItem.icon}
-						<div className="registrationWelcome__infoContent">
-							<h5>{infoItem.title}</h5>
-							<Text
-								text={infoItem.text}
-								type="infoLargeAlternative"
-							/>
-						</div>
-					</div>
-				))}
-			</div>
+			<ServiceExplanation className="registrationWelcome__explanation" />
 			<div className="registrationWelcome__buttonsWrapper">
 				<div>
 					<Text
